check_profile_picture_dims from userGather.py
This function is essentially a cleaning function, so it should remain here.
